mycadenza/requirements.txt
Babel==2.3.4
Django==1.10.5
django-phonenumber-field==1.1.0
phonenumberslite==8.0.1
postgres==2.2.1
psycopg2==2.6.2
pytz==2016.10
